requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.14.1=pypi_0
aiohttp=3.7.4.post0=py38h7f8727e_2
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=20.1.0=py38h27cfd23_1
async-timeout=3.0.1=py38h06a4308_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
audioread=2.1.9=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blessings=1.7=pypi_0
blinker=1.4=py38h06a4308_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h06a4308_1003
charset-normalizer=2.0.6=pypi_0
click=8.0.1=pyhd3eb1b0_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.8=py38hd23ed53_0
cycler=0.10.0=py38_0
cython=0.29.24=py38h295c915_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py38h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
flake8=3.9.2=pyhd3eb1b0_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
glib=2.69.1=h5202010_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.41.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.4.0=pypi_0
icu=58.2=he6710b0_3
idna=3.2=pyhd3eb1b0_0
importlib-metadata=4.8.1=py38h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.2.0=py38h06a4308_1
ipython=7.27.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.4=pyhd3eb1b0_0
jedi=0.18.0=py38h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
librosa=0.8.1=pypi_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.37.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.4.2=py38h06a4308_0
matplotlib-base=3.4.2=py38hab158f2_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
multidict=5.1.0=py38h27cfd23_2
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py38_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.3=py38h06a4308_0
numba=0.54.1=pypi_0
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.3=pypi_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.0.1=py38h06a4308_0
pooch=1.5.2=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyhca03da5_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.18.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py38h295c915_1
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
regex=2021.8.3=py38h7f8727e_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
resampy=0.2.2=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0=pypi_0
scipy=1.7.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.9.4=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torch=1.7.1+cu101=pypi_0
torch-tb-profiler=0.3.1=pypi_0
torchaudio=0.7.2=pypi_0
torchvision=0.8.2+cu101=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.62.2=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
typed-ast=1.4.3=py38h7f8727e_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
